# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Grama360 API
  version: 1.0.0
  description: Versioned REST API shared by the Flutter app and future channels.
servers:
  - url: https://api.example.invalid/api/v1
    description: Replace with the deployed API host.
security:
  - FirebaseBearer: []
paths:
...
  /me/favorites:
    get:
      summary: List the authenticated customer's saved providers.
      description: Returns only the caller's saved providers that are still active and publicly discoverable. Summaries never include contact numbers.
      parameters:
        - in: query
          name: language
          schema:
            type: string
            enum: [en, kn]
            default: en
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
        - in: query
          name: cursor
          description: Opaque nextCursor returned by the previous page; do not construct or edit it.
          schema:
            type: string
            maxLength: 256
      responses:
        '200':
          description: Paginated public-safe favorites page with an opaque nextCursor when more results remain.
          content:
            application/json:
              schema:
                type: object
                required: [items, hasMore, nextCursor, limit]
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      description: Public provider summary; no contact number.
                  hasMore:
                    type: boolean
                  nextCursor:
                    type: [string, 'null']
                    maxLength: 256
                  limit:
                    type: integer
        '400':
          description: Favorite page filters are invalid.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Customer role required or account disabled.
  /me/favorites/{providerId}:
    put:
      summary: Idempotently save a provider for the authenticated customer.
      description: The customer identity is derived from the Firebase session. Only active providers with active service categories can be saved.
      parameters:
        - in: path
          name: providerId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Favorite is saved; the response confirms the provider is saved.
        '400':
          description: Provider ID is invalid.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Customer role required or account disabled.
        '404':
          description: Provider is not active or publicly discoverable.
    delete:
      summary: Remove a provider from the authenticated customer's favorites.
      description: Idempotent; only the caller's favorite row is removed.
      parameters:
        - in: path
          name: providerId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Favorite is removed; the response confirms the provider is no longer saved.
        '400':
          description: Provider ID is invalid.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Customer role required or account disabled.
...
  /providers:
    get:
      summary: Search approved providers by service, text, and locality.
      description: Requires a signed-in account with the CUSTOMER role. Search and profile responses never contain a contact phone number.
      parameters:
        - in: query
          name: categoryId
          schema:
            type: string
            format: uuid
        - in: query
          name: q
          description: Provider, business, or service text; case-insensitive and supports English or Kannada.
          schema:
            type: string
            maxLength: 100
        - in: query
          name: location
          description: Matches a village/locality, taluk, or district in either stored language.
          schema:
            type: string
            maxLength: 100
        - in: query
          name: language
          schema:
            type: string
            enum: [en, kn]
            default: en
        - in: query
          name: availableNow
          description: If true, return only providers whose current availability is AVAILABLE.
          schema:
            type: boolean
            default: false
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            maximum: 100000
            default: 0
      responses:
        '200':
          description: Localized provider page with availability and aggregate review rating; no contact number.
        '400':
          description: One or more search filters are invalid.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Customer role required or account disabled.
  /providers/{providerId}:
    get:
      summary: Get an approved provider's public profile.
      description: Returns localized services, service area, availability, working hours and aggregate rating; never includes contact details.
      parameters:
        - in: path
          name: providerId
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: language
          schema:
            type: string
            enum: [en, kn]
            default: en
      responses:
        '200':
          description: Public-safe provider profile.
        '400':
          description: Provider ID or language is invalid.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Customer role required or account disabled.
        '404':
          description: Provider is not approved or available.
  /providers/{providerId}/call-intent:
    post:
      summary: Record a call-button tap and return the verified number for the Android dialer.
      description: Requires the CUSTOMER role. The API rechecks active provider status and returns the number only to this authenticated call action. A tap is not evidence that a call completed.
      parameters:
        - in: path
          name: providerId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Call intent recorded; contains a phone number for launching the native dialer.
        '400':
          description: Provider ID is invalid.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Customer role required or account disabled.
        '404':
          description: Provider is not active or available for contact.
...
components:
  securitySchemes:
    FirebaseBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID token; verified by the API with Firebase Admin.
...
  responses:
    Unauthorized:
      description: Missing, invalid, or expired Firebase ID token.
...